Move settings up gateways and cache to Setup
