+++ lib/rom/configuration.rb
@@ -7,8 +7,6 @@ require "rom/setup"
 require "rom/configuration_dsl"
 require "rom/support/configurable"
 require "rom/support/inflector"
-require "rom/gateway_registry"
-require "rom/relation_registry"
 
 module ROM
   class Configuration
@@ -28,29 +26,18 @@ module ROM
 
     NoDefaultAdapterError = Class.new(StandardError)
 
-    # @!attribute [r] gateways
-    #   @return [GatewayRegistry] Configured runtime gateways
-    attr_reader :gateways
-
     # @!attribute [r] setup
-    #   @return [Setup] Setup object which collects component classes and plugins
+    #   @return [Setup] Setup object which manages component and plugins
     attr_reader :setup
 
     # @!attribute [r] notifications
     #   @return [Notifications] Notification bus instance
     attr_reader :notifications
 
-    # @!attribute [r] cache
-    #   @return [Notifications] Cache
-    attr_reader :cache
-
-    # @!attribute [r] relations
-    #   @return [ROM::RelationRegistry] relations
-    attr_reader :relations
-
-    def_delegators :@setup, :register_relation, :register_command,
-                   :register_mapper, :register_plugin, :auto_register,
-                   :inflector, :inflector=, :components, :plugins
+    def_delegators :@setup, :cache, :relations, :gateways,
+      :register_relation, :register_command, :register_mapper,
+      :register_plugin, :auto_register, :inflector, :inflector=,
+      :components, :plugins
 
     # Initialize a new configuration
     #
@@ -60,15 +47,10 @@ module ROM
     def initialize(*args, &block)
       @notifications = Notifications.event_bus(:configuration)
 
-      @setup = Setup.new
-      @cache = Cache.new
-
       config.gateways = Config.new
-      @gateways = GatewayRegistry.new({}, cache: cache, config: config.gateways)
+      @setup = Setup.new(config: config.gateways)
 
       configure(*args, &block)
-
-      @relations = RelationRegistry.build
     end
 
     # @api public
@@ -90,7 +72,7 @@ module ROM
         end
       end
 
-      load_gateways
+      setup.load_gateways
 
       yield(self) if block_given?
 
@@ -165,24 +147,6 @@ module ROM
     end
 
     private
-
-    # @api private
-    def load_gateways
-      config.gateways.each do |name, gateway_config|
-        gateway =
-          if gateway_config.adapter.is_a?(Gateway)
-            gateway_config.adapter
-          else
-            Gateway.setup(gateway_config.adapter, gateway_config)
-          end
-
-        # TODO: this is here to keep backward compatibility
-        gateway_config.name = name
-        gateway.instance_variable_set(:"@config", gateway_config)
-
-        gateways.add(name, gateway)
-      end
-    end
 
     # @api private
     def load_config(config, hash)
