+++ lib/rom/setup.rb
@@ -1,7 +1,11 @@
 # frozen_string_literal: true
 
+require "rom/support/configurable"
 require "rom/loader"
 require "rom/components"
+require "rom/gateway"
+require "rom/gateway_registry"
+require "rom/relation_registry"
 
 module ROM
   # Setup objects collect component classes during setup/finalization process
@@ -18,10 +22,31 @@ module ROM
     attr_reader :components
 
     # @api private
-    def initialize(components: Components::Registry.new, auto_register: EMPTY_HASH)
+    attr_reader :gateways
+
+    # @api private
+    attr_reader :relations
+
+    # @api private
+    attr_reader :config
+
+    # @api private
+    attr_reader :cache
+
+    # @api private
+    def initialize(
+      components: Components::Registry.new,
+      auto_register: EMPTY_HASH,
+      cache: Cache.new,
+      config: Configurable::Config.new
+    )
       @plugins = []
+      @cache = cache
+      @config = config
       @inflector = Inflector
       @components = components
+      @gateways = GatewayRegistry.new({}, cache: cache, config: config)
+      @relations = RelationRegistry.build
       @auto_register = auto_register.merge(root_directory: nil, components: components)
     end
 
@@ -78,6 +103,24 @@ module ROM
     # @api private
     def register_plugin(plugin)
       plugins << plugin
+    end
+
+    # @api private
+    def load_gateways
+      config.each do |name, gateway_config|
+        gateway =
+          if gateway_config.adapter.is_a?(Gateway)
+            gateway_config.adapter
+          else
+            Gateway.setup(gateway_config.adapter, gateway_config)
+          end
+
+        # TODO: this is here to keep backward compatibility
+        gateway_config.name = name
+        gateway.instance_variable_set(:"@config", gateway_config)
+
+        gateways.add(name, gateway)
+      end
     end
 
     private
